requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==2.3.7
requests==2.32.3
google-generativeai==0.3.1
tqdm==4.67.1
watchdog==3.0.0 